+++ internal/retry/check.go
@@ -1,6 +1,7 @@
 package retry
 
 import (
+	"context"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/backoff"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/operation"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
@@ -14,12 +15,15 @@ func Check(err error) (
 	deleteSession bool,
 ) {
 	var e xerrors.Error
-	if xerrors.As(err, &e) {
-		return int64(e.Code()),
-			e.OperationStatus(),
-			e.BackoffType(),
-			e.MustDeleteSession()
+	switch {
+	case xerrors.As(err, &e):
+		return int64(e.Code()), e.OperationStatus(), e.BackoffType(), e.MustDeleteSession()
+
+	// context errors with explicit delete session
+	case xerrors.Is(err, context.Canceled), xerrors.Is(err, context.DeadlineExceeded):
+		return int64(-1), operation.Finished, backoff.TypeNoBackoff, true
 	}
+
 	return -1,
 		operation.Finished, // it's finished, not need any retry attempts
 		backoff.TypeNoBackoff,
@@ -28,10 +32,17 @@ func Check(err error) (
 
 func MustDeleteSession(err error) bool {
 	var e xerrors.Error
-	if xerrors.As(err, &e) {
+	switch {
+	case xerrors.As(err, &e):
 		return e.MustDeleteSession()
+
+	// context errors with explicit delete session
+	case xerrors.Is(err, context.Canceled), xerrors.Is(err, context.DeadlineExceeded):
+		return true
+
+	default:
+		return false
 	}
-	return false
 }
 
 func MustRetry(err error, isOperationIdempotent bool) bool {