+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added log topic api reader for internal logger
+
 ## v3.37.2
 * Fixed nil pointer exception in topic reader if reconnect failed