+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Marked the context errors as required to delete session
 * Added log topic api reader for internal logger
 
 ## v3.37.2